# src/handlers/openapi.yaml
openapi: "3.0.3"
info:
  title: Character Gallery API
  description: |
    A RESTful API built in Go to create, manage and see a gallery of Role
    Playing Game characters. Supports full CRUD operations, including base
    stats and character appearance customization, plus a turn-based tactical
    RPG combat system.
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Current server (adapts automatically to localhost or Cloudflare)

# All endpoints require a valid JWT by default. Public endpoints
# (register, login, metrics, docs) override this with `security: []`.
security:
  - bearerAuth: []
paths:
...
  /items:
    post:
      operationId: createItem
      summary: Create an item
      description: |
        Creates a new item and adds it to the item pool. **Admin only**;
        requires a JWT with role `admin`.
      tags: [Items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: getItemPool
      summary: Get the item pool
      description: Returns a paginated list of all items in the item pool.
      tags: [Items]
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page (minimum 20)
          schema:
            type: integer
            minimum: 20
            default: 20
      responses:
        "200":
          description: Paginated list of items
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Item"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT returned by `/auth/login`. Include as `Authorization: Bearer <token>`.
        Send on every authenticated request. The token embeds the user's role
        (`user` or `admin`); `admin` can modify/delete any character.
  schemas:
...
    Item:
      type: object
      description: An item from the item pool.
      required: [id, name, type, description, equippable, rarity]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          minLength: 3
          maxLength: 50
          example: "Master Sword"
        type:
          type: string
          enum:
            [armor, ring, weapon, shield, tool, adventuring_gear, rod,
             staff, wand, scroll, potion, ammo, consumable, wondrous_item]
          example: "weapon"
        description:
          type: string
          minLength: 3
          maxLength: 300
          example: "A legendary sword with immense power."
        equippable:
          type: boolean
          example: true
        rarity:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        damage:
          type: integer
          minimum: 0
          example: 34
        defense:
          type: integer
          minimum: 0
          example: 23
        heal_amount:
          type: integer
          minimum: 0
          example: 10
        mana_cost:
          type: integer
          minimum: 0
          example: 4
        duration:
          type: integer
          minimum: 0
          example: 233
        cooldown:
          type: integer
          minimum: 0
          example: 120
        capacity:
          type: integer
          minimum: 0
          example: 3
    CreateItemRequest:
      type: object
      description: Payload for creating an item (no id).
      required: [name, type, description, equippable, rarity]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          example: "Master Sword"
        type:
          type: string
          enum:
            [armor, ring, weapon, shield, tool, adventuring_gear, rod,
             staff, wand, scroll, potion, ammo, consumable, wondrous_item]
          example: "weapon"
        description:
          type: string
          minLength: 3
          maxLength: 300
          example: "A legendary sword with immense power."
        equippable:
          type: boolean
          example: true
        rarity:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        damage:
          type: integer
          minimum: 0
          example: 34
        defense:
          type: integer
          minimum: 0
          example: 23
        heal_amount:
          type: integer
          minimum: 0
          example: 10
        mana_cost:
          type: integer
          minimum: 0
          example: 4
        duration:
          type: integer
          minimum: 0
          example: 233
        cooldown:
          type: integer
          minimum: 0
          example: 120
        capacity:
          type: integer
          minimum: 0
          example: 3
...
    Pagination:
      type: object
      description: Pagination metadata.
      required: [page, limit, total, has_next]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        has_next:
          type: boolean
          example: true
    ApiError:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          example: "Invalid ID"
        code:
          type: string
          example: "BAD_REQUEST"
        details:
          type: object
          description: Additional error context (may vary per error)
          example: { "id": "abc" }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Not the character owner (and not an admin)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        text/plain:
          schema:
            type: string
            example: "429 too many requests"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"