# src/internal/postgres/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Character struct {
	ID        int64
	OwnerID   uuid.UUID
	Name      string
	BodyType  string
	Species   string
	Class     string
	Level     int32
	Xp        int32
	HpMax     int32
	HpCurrent int32
}

type Customization struct {
	CharacterID int64
	Hair        int16
	Face        int16
	Shirt       int16
	Pants       int16
	Shoes       int16
}

type Inventory struct {
	CharacterID int64
	ItemID      int64
	Quantity    int32
	IsEquipped  bool
}

type Item struct {
	ID          int64
	Name        string
	Type        string
	Description string
	Equippable  bool
	Rarity      int32
	Damage      pgtype.Int4
	Defense     pgtype.Int4
	HealAmount  pgtype.Int4
	ManaCost    pgtype.Int4
	Duration    pgtype.Int4
	Cooldown    pgtype.Int4
	Capacity    pgtype.Int4
}

type Stat struct {
	CharacterID  int64
	Strength     int16
	Dexterity    int16
	Constitution int16
	Intelligence int16
	Wisdom       int16
	Charisma     int16
}

type User struct {
	ID           uuid.UUID
	Username     string
	PasswordHash string
	Role         string
	CreatedAt    pgtype.Timestamptz
}